Infra/3_Applications/data-fetcher/requirements.txt
azure-servicebus==7.12.3
jmespath==1.0.1
PyYAML==6.0.1
${COMMON_PACKAGE_PATH}
